# openapi.yaml
openapi: 3.1.0
info:
  title: Inventory API
  version: 1.0.0
servers:
- description: Generated server url
  url: http://localhost:8080
paths:
  /inventory:
    get:
      description: "Returns SKUs sorted by skuId, at most 250 per response. When more\
        \ SKUs follow, the Link header holds the next page's URL; follow it until\
        \ a response has no Link to list all inventory. If no SKUs exist, returns\
        \ an empty array."
      operationId: listInventory
      parameters:
      - description: "Optional page size, 1 to 250 (default 250). Larger values mean\
          \ 250; other values are ignored and the default applies."
        in: query
        name: limit
        required: false
        schema:
          type: integer
          default: 250
          maximum: 250
          minimum: 1
      - description: "Optional cursor: return only SKUs whose skuId sorts after this\
          \ value, up to the page size. It must not be repeated."
        in: query
        name: after
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
          description: "One page of inventory items, at most 250, sorted by skuId"
          headers:
            Link:
              description: "Next page, when more SKUs follow: <URL>; rel=\"next\""
              schema:
                type: string
              style: simple
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: "Invalid request: the query string can't be decoded or repeats\
            \ after"
      summary: List all inventory
      tags:
      - inventory
  /inventory/{skuId}:
    get:
      operationId: getInventory
      parameters:
      - description: "SKU ID: 1 to 64 characters; letters, digits, '.', '_' or '-',\
          \ starting with a letter or digit. Case-sensitive."
        in: path
        name: skuId
        required: true
        schema:
          type: string
          maxLength: 64
          minLength: 1
          pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
          description: Current inventory state for the sku
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: SKU not found
      summary: Get inventory for a SKU
      tags:
      - inventory
    post:
      operationId: createInventory
      parameters:
      - description: "SKU ID: 1 to 64 characters; letters, digits, '.', '_' or '-',\
          \ starting with a letter or digit. Case-sensitive."
        in: path
        name: skuId
        required: true
        schema:
          type: string
          maxLength: 64
          minLength: 1
          pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"
      - description: "Optional UUID. The same key with the same request replays the\
          \ first response. A different request, or a key older than 24h, returns\
          \ 400."
        in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryQuantity"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
          description: Current state of the item after update
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Invalid request
      summary: Create or update inventory for a SKU
      tags:
      - inventory
  /inventory/{skuId}/purchase:
    post:
      operationId: purchaseItem
      parameters:
      - description: "SKU ID: 1 to 64 characters; letters, digits, '.', '_' or '-',\
          \ starting with a letter or digit. Case-sensitive."
        in: path
        name: skuId
        required: true
        schema:
          type: string
          maxLength: 64
          minLength: 1
          pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"
      - description: "Optional UUID. The same key with the same request replays the\
          \ first response. A different request, or a key older than 24h, returns\
          \ 400."
        in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryQuantity"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
          description: Purchase successful; remaining inventory for the item
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Insufficient inventory or invalid request
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: SKU not found
      summary: Purchase a quantity of a SKU
      tags:
      - inventory
  /v2/inventory:
    get:
      description: "Returns SKUs sorted by skuId with their quantity and details (absent\
        \ for a SKU that has none), at most 250 per response; the same limit, after\
        \ and Link rules as GET /inventory."
      operationId: listSkus
      parameters:
      - description: "Optional page size, 1 to 250 (default 250). Larger values mean\
          \ 250; other values are ignored and the default applies."
        in: query
        name: limit
        required: false
        schema:
          type: integer
          default: 250
          maximum: 250
          minimum: 1
      - description: "Optional cursor: return only SKUs whose skuId sorts after this\
          \ value, up to the page size. It must not be repeated."
        in: query
        name: after
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SkuItem"
          description: "One page of SKUs with details, at most 250, sorted by skuId"
          headers:
            Link:
              description: "Next page, when more SKUs follow: <URL>; rel=\"next\""
              schema:
                type: string
              style: simple
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: "Invalid request: the query string can't be decoded or repeats\
            \ after"
      summary: List all SKUs with details
      tags:
      - inventory-v2
  /v2/inventory/{skuId}:
    get:
      operationId: getSku
      parameters:
      - description: "SKU ID: 1 to 64 characters; letters, digits, '.', '_' or '-',\
          \ starting with a letter or digit. Case-sensitive."
        in: path
        name: skuId
        required: true
        schema:
          type: string
          maxLength: 64
          minLength: 1
          pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkuItem"
          description: The SKU with its quantity and details
          headers:
            ETag:
              description: "The details version, a strong validator for If-Match;\
                \ \"0\" before any details"
              schema:
                type: string
              style: simple
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: SKU not found
      summary: Get a SKU with its details
      tags:
      - inventory-v2
    post:
      description: "Adds stock, creating the SKU if needed; like POST /inventory/{skuId},\
        \ but the Idempotency-Key is required and the response is a SkuItem. The same\
        \ key with the same request replays the first response."
      operationId: addStock
      parameters:
      - description: "SKU ID: 1 to 64 characters; letters, digits, '.', '_' or '-',\
          \ starting with a letter or digit. Case-sensitive."
        in: path
        name: skuId
        required: true
        schema:
          type: string
          maxLength: 64
          minLength: 1
          pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"
      - description: "Required UUID. The same key with the same request replays the\
          \ first response. A missing or malformed key, a different request, or a\
          \ key older than 24h returns 400."
        in: header
        name: Idempotency-Key
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryQuantity"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkuItem"
          description: The SKU after the update
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Invalid request
      summary: Add stock to a SKU
      tags:
      - inventory-v2
  /v2/inventory/{skuId}/details:
    put:
      description: "Replaces the whole SkuDetails of the SKU. A SKU that does not\
        \ exist is created with quantity 0 (201, ETag \"1\"); an existing SKU keeps\
        \ its stock (200, new ETag). If-Match (strong ETags, or \"*\" for any existing\
        \ SKU) and If-None-Match: * (create only) are optional preconditions; without\
        \ one the PUT is unconditional. Any If-Match on a SKU that does not exist\
        \ is 412. Idempotency-Key is ignored: PUT is idempotent by method."
      operationId: putSkuDetails
      parameters:
      - description: "Optional strong ETag(s) from a previous response, or \"*\" for\
          \ any existing SKU; 412 when none matches or the SKU does not exist"
        in: header
        name: If-Match
        schema:
          type: string
      - description: "Optional; only \"*\" is supported: create only, 412 when the\
          \ SKU already exists. Any other value is 400"
        in: header
        name: If-None-Match
        schema:
          type: string
      - description: "SKU ID: 1 to 64 characters; letters, digits, '.', '_' or '-',\
          \ starting with a letter or digit. Case-sensitive."
        in: path
        name: skuId
        required: true
        schema:
          type: string
          maxLength: 64
          minLength: 1
          pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkuDetails"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkuItem"
          description: The SKU with its replaced details
          headers:
            ETag:
              description: "The details version, a strong validator for If-Match;\
                \ \"0\" before any details"
              schema:
                type: string
              style: simple
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkuItem"
          description: The SKU was created at quantity 0 with these details
          headers:
            ETag:
              description: "The details version of a new SKU, always \"1\""
              schema:
                type: string
              style: simple
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Invalid request
        "412":
          content:
            text/plain:
              schema:
                type: string
          description: Details changed since you read them. Reload the SKU and retry
            with its new ETag.
      summary: Create a SKU or replace its details
      tags:
      - inventory-v2
  /v2/inventory/{skuId}/purchase:
    post:
      description: "Like POST /inventory/{skuId}/purchase, but the Idempotency-Key\
        \ is required and the response is a SkuItem. The same key with the same request\
        \ replays the first response."
      operationId: purchaseStock
      parameters:
      - description: "SKU ID: 1 to 64 characters; letters, digits, '.', '_' or '-',\
          \ starting with a letter or digit. Case-sensitive."
        in: path
        name: skuId
        required: true
        schema:
          type: string
          maxLength: 64
          minLength: 1
          pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"
      - description: "Required UUID. The same key with the same request replays the\
          \ first response. A missing or malformed key, a different request, or a\
          \ key older than 24h returns 400."
        in: header
        name: Idempotency-Key
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryQuantity"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkuItem"
          description: Purchase successful; the SKU with its remaining quantity
        "400":
          content:
            text/plain:
              schema:
                type: string
          description: Insufficient inventory or invalid request
        "404":
          content:
            text/plain:
              schema:
                type: string
          description: SKU not found
      summary: Purchase a quantity of a SKU
      tags:
      - inventory-v2
components:
  schemas:
    InventoryItem:
      type: object
      properties:
        quantity:
          type: integer
          format: int64
          minimum: 0
        skuId:
          type: string
    InventoryQuantity:
      type: object
      properties:
        quantity:
          type: integer
          format: int32
          minimum: 1
      required:
      - quantity
    SkuCost:
      type: object
      description: A cost in minor units of a currency
      properties:
        amount:
          type: integer
          format: int64
          description: "Minor units (e.g. cents), never negative"
          minimum: 0
        currency:
          type: string
          description: Three-letter uppercase currency code
          pattern: "^[A-Z]{3}$"
      required:
      - amount
      - currency
    SkuDetails:
      type: object
      properties:
        cost:
          $ref: "#/components/schemas/SkuCost"
        description:
          type: string
          description: Up to 2000 characters (UTF-16 units); newlines and tabs allowed;
            default ""
          maxLength: 2000
        images:
          type: array
          description: "Up to 10 absolute http or https URLs in ASCII (percent-encoded);\
            \ default []"
          items:
            type: string
            format: uri
            maxLength: 2048
          maxItems: 10
        name:
          type: string
          description: "1 to 120 characters (UTF-16 units), not blank, no control\
            \ characters"
          maxLength: 120
          minLength: 1
      required:
      - name
    SkuItem:
      type: object
      properties:
        details:
          $ref: "#/components/schemas/SkuDetails"
        quantity:
          type: integer
          format: int64
          minimum: 0
        skuId:
          type: string
      required:
      - quantity
      - skuId